chore: rename instrument to just sentry
